+++ test-clients/src/main/java/com/hedera/services/bdd/suites/SuiteRunner.java
@@ -75,23 +75,23 @@ import com.hedera.services.bdd.suites.misc.MixedOpsTransactionsSuite;
 import com.hedera.services.bdd.suites.misc.OneOfEveryTransaction;
 import com.hedera.services.bdd.suites.misc.ZeroStakeNodeTest;
 import com.hedera.services.bdd.suites.perf.AccountBalancesClientSaveLoadTest;
-import com.hedera.services.bdd.suites.perf.ContractCallLoadTest;
-import com.hedera.services.bdd.suites.perf.CreateTopicPerfSuite;
-import com.hedera.services.bdd.suites.perf.CryptoCreatePerfSuite;
-import com.hedera.services.bdd.suites.perf.CryptoTransferLoadTest;
+import com.hedera.services.bdd.suites.perf.contract.ContractCallLoadTest;
+import com.hedera.services.bdd.suites.perf.topic.CreateTopicPerfSuite;
+import com.hedera.services.bdd.suites.perf.crypto.CryptoCreatePerfSuite;
+import com.hedera.services.bdd.suites.perf.crypto.CryptoTransferLoadTest;
 import com.hedera.services.bdd.suites.perf.FileContractMemoPerfSuite;
-import com.hedera.services.bdd.suites.perf.FileUpdateLoadTest;
-import com.hedera.services.bdd.suites.perf.HCSChunkingRealisticPerfSuite;
-import com.hedera.services.bdd.suites.perf.MixedOpsLoadTest;
-import com.hedera.services.bdd.suites.perf.MixedOpsMemoPerfSuite;
-import com.hedera.services.bdd.suites.perf.MixedTransferAndSubmitLoadTest;
-import com.hedera.services.bdd.suites.perf.MixedTransferCallAndSubmitLoadTest;
+import com.hedera.services.bdd.suites.perf.file.FileUpdateLoadTest;
+import com.hedera.services.bdd.suites.perf.topic.HCSChunkingRealisticPerfSuite;
+import com.hedera.services.bdd.suites.perf.mixedops.MixedOpsLoadTest;
+import com.hedera.services.bdd.suites.perf.mixedops.MixedOpsMemoPerfSuite;
+import com.hedera.services.bdd.suites.perf.mixedops.MixedTransferAndSubmitLoadTest;
+import com.hedera.services.bdd.suites.perf.mixedops.MixedTransferCallAndSubmitLoadTest;
 import com.hedera.services.bdd.suites.perf.QueryOnlyLoadTest;
-import com.hedera.services.bdd.suites.perf.ReadyToRunScheduledXfersLoad;
-import com.hedera.services.bdd.suites.perf.SubmitMessageLoadTest;
-import com.hedera.services.bdd.suites.perf.TokenRelStatusChanges;
-import com.hedera.services.bdd.suites.perf.TokenTransferBasicLoadTest;
-import com.hedera.services.bdd.suites.perf.TokenTransfersLoadProvider;
+import com.hedera.services.bdd.suites.perf.schedule.ReadyToRunScheduledXfersLoad;
+import com.hedera.services.bdd.suites.perf.topic.SubmitMessageLoadTest;
+import com.hedera.services.bdd.suites.perf.token.TokenRelStatusChanges;
+import com.hedera.services.bdd.suites.perf.token.TokenTransferBasicLoadTest;
+import com.hedera.services.bdd.suites.perf.token.TokenTransfersLoadProvider;
 import com.hedera.services.bdd.suites.reconnect.CheckUnavailableNode;
 import com.hedera.services.bdd.suites.reconnect.CreateAccountsBeforeReconnect;
 import com.hedera.services.bdd.suites.reconnect.CreateFilesBeforeReconnect;
