+++ test-clients/src/main/java/com/hedera/services/bdd/suites/perf/contract/ContractCallLoadTest.java
@@ -1,4 +1,4 @@
-package com.hedera.services.bdd.suites.perf;
+package com.hedera.services.bdd.suites.perf.contract;
 
 /*-
  * ‌
@@ -9,9 +9,9 @@ package com.hedera.services.bdd.suites.perf;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * 
  *      http://www.apache.org/licenses/LICENSE-2.0
- *
+ * 
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -24,44 +24,34 @@ import com.hedera.services.bdd.spec.HapiApiSpec;
 import com.hedera.services.bdd.spec.HapiSpecOperation;
 import com.hedera.services.bdd.spec.infrastructure.meta.ContractResources;
 import com.hedera.services.bdd.spec.utilops.LoadTest;
+import com.hedera.services.bdd.suites.perf.PerfTestLoadSettings;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Supplier;
+import java.util.stream.IntStream;
 
 import static com.hedera.services.bdd.spec.HapiApiSpec.defaultHapiSpec;
 import static com.hedera.services.bdd.spec.queries.QueryVerbs.getContractInfo;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.contractCall;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.contractCreate;
-import static com.hedera.services.bdd.spec.transactions.TxnVerbs.cryptoCreate;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.fileCreate;
+import static com.hedera.services.bdd.spec.utilops.UtilVerbs.inParallel;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.logIt;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.withOpContext;
-import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.BUSY;
-import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.DUPLICATE_TRANSACTION;
-import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.PLATFORM_TRANSACTION_NOT_CREATED;
-import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.SUCCESS;
-import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.UNKNOWN;
 
-public class ContractBigArrayLoadTest extends LoadTest {
-	private static final Logger log = LogManager.getLogger(ContractBigArrayLoadTest.class);
-	private static int sizeInKb = 4;
+public class ContractCallLoadTest extends LoadTest {
+	private static final Logger log = LogManager.getLogger(ContractCallLoadTest.class);
 
 	public static void main(String... args) {
-		int usedArgs = parseArgs(args);
-
-		// parsing local argument specific to this test
-		if (args.length > usedArgs) {
-			sizeInKb = Integer.parseInt(args[usedArgs]);
-			log.info("Set sizeInKb as " + sizeInKb);
-		}
+		parseArgs(args);
 
 		/* Has a static initializer whose behavior seems influenced by initialization of ForkJoinPool#commonPool. */
 		new org.ethereum.crypto.HashUtil();
 
-		ContractBigArrayLoadTest suite = new ContractBigArrayLoadTest();
+		ContractCallLoadTest suite = new ContractCallLoadTest();
 		suite.setReportStats(true);
 		suite.runSuiteSync();
 	}
@@ -79,15 +69,21 @@ public class ContractBigArrayLoadTest extends LoadTest {
 	private HapiApiSpec runContractCalls() {
 		PerfTestLoadSettings settings = new PerfTestLoadSettings();
 		final AtomicInteger submittedSoFar = new AtomicInteger(0);
-		long setValue = 0x1234abdeL;
+		final String DEPOSIT_MEMO = "So we out-danced thought, body perfection brought...";
+
 		Supplier<HapiSpecOperation[]> callBurst = () -> new HapiSpecOperation[] {
-				contractCall("perf", ContractResources.BIG_ARRAY_CHANGE_ARRAY_ABI, setValue)
-						.noLogging()
-						.payingWith("sender")
-						.suppressStats(true)
-						.hasKnownStatusFrom(UNKNOWN, SUCCESS)
-						.hasRetryPrecheckFrom(BUSY, DUPLICATE_TRANSACTION, PLATFORM_TRANSACTION_NOT_CREATED)
-						.deferStatusResolution()
+				inParallel(IntStream.range(0, settings.getBurstSize())
+						.mapToObj(i ->
+								contractCall("perf", ContractResources.VERBOSE_DEPOSIT_ABI, i + 1, 0, DEPOSIT_MEMO)
+										.sending(i + 1)
+										.noLogging()
+										.suppressStats(true)
+										.deferStatusResolution())
+						.toArray(n -> new HapiSpecOperation[n])),
+				logIt(ignore ->
+						String.format(
+								"Now a total of %d transactions submitted.",
+								submittedSoFar.addAndGet(settings.getBurstSize()))),
 		};
 
 		return defaultHapiSpec("runContractCalls")
@@ -95,20 +91,11 @@ public class ContractBigArrayLoadTest extends LoadTest {
 						withOpContext((spec, ignore) -> settings.setFrom(spec.setup().ciPropertiesMap())),
 						logIt(ignore -> settings.toString())
 				).when(
-						cryptoCreate("sender").balance(initialBalance.getAsLong())
-								.withRecharging()
-								.rechargeWindow(3)
-								.hasRetryPrecheckFrom(BUSY, DUPLICATE_TRANSACTION, PLATFORM_TRANSACTION_NOT_CREATED),
-						fileCreate("contractBytecode").path(ContractResources.BIG_ARRAY_BYTECODE_PATH)
-								.hasRetryPrecheckFrom(BUSY, DUPLICATE_TRANSACTION, PLATFORM_TRANSACTION_NOT_CREATED),
-						contractCreate("perf").bytecode("contractBytecode")
-								.hasRetryPrecheckFrom(BUSY, DUPLICATE_TRANSACTION, PLATFORM_TRANSACTION_NOT_CREATED),
-						getContractInfo("perf").hasExpectedInfo().logged(),
-
-						// Initialize storage size
-						contractCall("perf", ContractResources.BIG_ARRAY_SET_SIZE_IN_KB_ABI, sizeInKb)
-								.hasRetryPrecheckFrom(BUSY, DUPLICATE_TRANSACTION, PLATFORM_TRANSACTION_NOT_CREATED)
-								.gas(600000)
+						fileCreate("contractBytecode").path(ContractResources.VERBOSE_DEPOSIT_BYTECODE_PATH),
+						contractCreate("perf").bytecode("contractBytecode"),
+						fileCreate("lookupBytecode").path(ContractResources.BALANCE_LOOKUP_BYTECODE_PATH),
+						contractCreate("balanceLookup").bytecode("lookupBytecode").balance(1L),
+						getContractInfo("perf").hasExpectedInfo().logged()
 
 				).then(
 						defaultLoadTest(callBurst, settings)
@@ -120,3 +107,5 @@ public class ContractBigArrayLoadTest extends LoadTest {
 		return log;
 	}
 }
+
+
