+++ test-clients/src/main/java/com/hedera/services/bdd/suites/perf/contract/ContractCallLocalPerfSuite.java
@@ -1,4 +1,4 @@
-package com.hedera.services.bdd.suites.perf;
+package com.hedera.services.bdd.suites.perf.contract;
 
 /*-
  * ‌
@@ -23,7 +23,6 @@ package com.hedera.services.bdd.suites.perf;
 import com.hedera.services.bdd.spec.HapiApiSpec;
 
 import static com.hedera.services.bdd.spec.queries.QueryVerbs.contractCallLocal;
-import static com.hedera.services.bdd.spec.queries.QueryVerbs.getContractInfo;
 
 import com.hedera.services.bdd.spec.infrastructure.meta.ContractResources;
 import com.hedera.services.bdd.spec.utilops.UtilVerbs;
@@ -31,30 +30,26 @@ import com.hedera.services.bdd.suites.HapiApiSuite;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import java.math.BigInteger;
 import java.util.List;
 
 import static com.hedera.services.bdd.spec.HapiApiSpec.*;
-import static com.hedera.services.bdd.spec.transactions.TxnVerbs.contractCall;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.contractCreate;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.fileCreate;
-import static com.hedera.services.bdd.spec.assertions.ContractFnResultAsserts.resultWith;
-import static com.hedera.services.bdd.spec.assertions.ContractFnResultAsserts.isLiteralResult;
 
-public class ContractCallPerfSuite extends HapiApiSuite {
-	private static final Logger log = LogManager.getLogger(ContractCallPerfSuite.class);
+public class ContractCallLocalPerfSuite extends HapiApiSuite {
+	private static final Logger log = LogManager.getLogger(ContractCallLocalPerfSuite.class);
 
 	public static void main(String... args) {
 		/* Has a static initializer whose behavior seems influenced by initialization of ForkJoinPool#commonPool. */
 		new org.ethereum.crypto.HashUtil();
 
-		new ContractCallPerfSuite().runSuiteSync();
+		new ContractCallLocalPerfSuite().runSuiteSync();
 	}
 
 	@Override
 	protected List<HapiApiSpec> getSpecsInSuite() {
 		return List.of(
-				contractCallPerf()
+				contractCallLocalPerf()
 		);
 	}
 
@@ -68,42 +63,30 @@ public class ContractCallPerfSuite extends HapiApiSuite {
 		return true;
 	}
 
-	HapiApiSpec contractCallPerf() {
+	HapiApiSpec contractCallLocalPerf() {
 		final int NUM_CALLS = 1_000;
-		final long ENDING_BALANCE = NUM_CALLS * (NUM_CALLS + 1)	 / 2;
-		final String DEPOSIT_MEMO = "So we out-danced thought, body perfection brought...";
 
-		return defaultHapiSpec("ContractCallPerf")
+		return defaultHapiSpec("ContractCallLocalPerf")
 				.given(
-						fileCreate("contractBytecode").path(ContractResources.VERBOSE_DEPOSIT_BYTECODE_PATH),
-						contractCreate("perf").bytecode("contractBytecode"),
-						fileCreate("lookupBytecode").path(ContractResources.BALANCE_LOOKUP_BYTECODE_PATH),
-						contractCreate("balanceLookup").bytecode("lookupBytecode").balance(1L)
+						fileCreate("bytecode").path(ContractResources.BALANCE_LOOKUP_BYTECODE_PATH),
+						contractCreate("contract").bytecode("bytecode").balance(1_000L)
 				).when(
-						getContractInfo("perf").hasExpectedInfo().logged(),
-						UtilVerbs.startThroughputObs("contractCall").msToSaturateQueues(50L)
+						contractCallLocal(
+								"contract",
+								ContractResources.BALANCE_LOOKUP_ABI,
+								spec -> new Object[] {
+										spec.registry().getContractId("contract").getContractNum()
+								}).recordNodePaymentAs("cost"),
+						UtilVerbs.startThroughputObs("contractCallLocal")
 				).then(
-						UtilVerbs.inParallel(asOpArray(NUM_CALLS, i ->
-								contractCall("perf", ContractResources.VERBOSE_DEPOSIT_ABI, i + 1, 0, DEPOSIT_MEMO)
-										.sending(i + 1)
-										.deferStatusResolution())),
-						UtilVerbs.finishThroughputObs("contractCall")
-								.gatedByQuery(() ->
+						UtilVerbs.inParallel(asOpArray(NUM_CALLS, ignore ->
 										contractCallLocal(
-												"balanceLookup",
+												"contract",
 												ContractResources.BALANCE_LOOKUP_ABI,
 												spec -> new Object[] {
-													spec.registry().getContractId("perf").getContractNum()
-												}
-										).has(
-												resultWith().resultThruAbi(
-														ContractResources.BALANCE_LOOKUP_ABI,
-														isLiteralResult(
-																new Object[] { BigInteger.valueOf(ENDING_BALANCE) }
-														)
-												)
-										).noLogging()
-								)
+														spec.registry().getContractId("contract").getContractNum()
+												}).nodePayment(spec -> spec.registry().getAmount("cost")))),
+						UtilVerbs.finishThroughputObs("contractCallLocal")
 				);
 	}
