+++ test-clients/src/main/java/com/hedera/services/bdd/suites/perf/crypto/CryptoCreatePerfSuite.java
@@ -1,4 +1,4 @@
-package com.hedera.services.bdd.suites.perf;
+package com.hedera.services.bdd.suites.perf.crypto;
 
 /*-
  * ‌
@@ -21,9 +21,7 @@ package com.hedera.services.bdd.suites.perf;
  */
 
 import com.hedera.services.bdd.spec.HapiApiSpec;
-import com.hedera.services.bdd.spec.transactions.TxnUtils;
 import com.hedera.services.bdd.spec.utilops.LoadTest;
-import com.hedera.services.bdd.spec.utilops.UtilVerbs;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -31,23 +29,21 @@ import java.util.List;
 
 import static com.hedera.services.bdd.spec.HapiApiSpec.defaultHapiSpec;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.cryptoCreate;
-import static com.hedera.services.bdd.spec.transactions.TxnVerbs.tokenCreate;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.freeze;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.inParallel;
 
-public class TokenCreatePerfSuite extends LoadTest {
-	private static final Logger log = LogManager.getLogger(TokenCreatePerfSuite.class);
+public class CryptoCreatePerfSuite extends LoadTest {
+	private static final Logger log = LogManager.getLogger(CryptoCreatePerfSuite.class);
 
 	public static void main(String... args) {
-		parseArgs(args);
-		TokenCreatePerfSuite suite = new TokenCreatePerfSuite();
+		CryptoCreatePerfSuite suite = new CryptoCreatePerfSuite();
 		suite.setReportStats(true);
 		suite.runSuiteSync();
 	}
 
 	@Override
 	protected List<HapiApiSpec> getSpecsInSuite() {
-		return List.of(runTokenCreates());
+		return List.of(runCryptoCreates());
 	}
 
 	@Override
@@ -55,26 +51,29 @@ public class TokenCreatePerfSuite extends LoadTest {
 		return true;
 	}
 
-	private HapiApiSpec runTokenCreates() {
-		final int NUM_CREATES = 100000;
-		return defaultHapiSpec("tokenCreatePerf")
+	private HapiApiSpec runCryptoCreates() {
+		final int NUM_CREATES = 1000000;
+		return defaultHapiSpec("cryptoCreatePerf")
 				.given(
 				).when(
 						inParallel(
 								asOpArray(NUM_CREATES, i ->
-										(i == (NUM_CREATES - 1)) ? tokenCreate("testToken" + i)
-												.payingWith(GENESIS)
-												.initialSupply(100_000_000_000L)
-												.signedBy(GENESIS):
-												tokenCreate("testToken" + i)
+										(i == (NUM_CREATES - 1)) ? cryptoCreate("testAccount" + i)
+												.balance(100_000_000_000L)
+												.key(GENESIS)
+												.withRecharging()
+												.rechargeWindow(30)
+												.payingWith(GENESIS) :
+												cryptoCreate("testAccount" + i)
+														.balance(100_000_000_000L)
+														.key(GENESIS)
+														.withRecharging()
+														.rechargeWindow(30)
 														.payingWith(GENESIS)
-														.signedBy(GENESIS)
-														.initialSupply(100_000_000_000L)
 														.deferStatusResolution()
 								)
 						)
 				).then(
-						UtilVerbs.sleepFor(200000),
 						freeze().payingWith(GENESIS).startingIn(60).seconds().andLasting(1).minutes()
 				);
 	}
