+++ test-clients/src/main/java/com/hedera/services/bdd/suites/perf/token/AdHocTokenTransfers.java
@@ -1,4 +1,4 @@
-package com.hedera.services.bdd.suites.perf;
+package com.hedera.services.bdd.suites.perf.token;
 
 /*-
  * ‌
