+++ test-clients/src/main/java/com/hedera/services/bdd/suites/perf/token/TokenRelStatusChanges.java
@@ -1,4 +1,4 @@
-package com.hedera.services.bdd.suites.perf;
+package com.hedera.services.bdd.suites.perf.token;
 
 /*-
  * ‌
